+++ tests/serialization/test_pydantic_models.py
@@ -33,7 +33,7 @@ from airflow.models.asset import (
     DagScheduleAssetReference,
     TaskOutletAssetReference,
 )
-from airflow.models.dag import DAG, DagModel, create_timetable
+from airflow.models.dag import DAG, DagModel
 from airflow.serialization.pydantic.asset import AssetEventPydantic
 from airflow.serialization.pydantic.dag import DagModelPydantic
 from airflow.serialization.pydantic.dag_run import DagRunPydantic
@@ -180,12 +180,11 @@ def test_serializing_pydantic_dagrun(session, create_task_instance):
     ],
 )
 def test_serializing_pydantic_dagmodel(schedule):
-    timetable = create_timetable(schedule, timezone.utc)
     dag_model = DagModel(
         dag_id="test-dag",
         fileloc="/tmp/dag_1.py",
-        timetable_summary=timetable.summary,
-        timetable_description=timetable.description,
+        timetable_summary="summary",
+        timetable_description="desc",
         is_active=True,
         is_paused=False,
     )
@@ -196,8 +195,8 @@ def test_serializing_pydantic_dagmodel(schedule):
     deserialized_model = DagModelPydantic.model_validate_json(json_string)
     assert deserialized_model.dag_id == "test-dag"
     assert deserialized_model.fileloc == "/tmp/dag_1.py"
-    assert deserialized_model.timetable_summary == timetable.summary
-    assert deserialized_model.timetable_description == timetable.description
+    assert deserialized_model.timetable_summary == "summary"
+    assert deserialized_model.timetable_description == "desc"
     assert deserialized_model.is_active is True
     assert deserialized_model.is_paused is False
